Enhance Sign Handling

+++ src/Crypto/Hash/CRC16R_CPP.cpp
@@ -14,11 +14,11 @@ extern "C" void CRC16R_InitTable(UInt16 *tab, UInt16 polynomial)
 		{
 			if (v & 1)
 			{
-				v = (v >> 1) ^ polynomial;
+				v = (UInt16)(v >> 1) ^ polynomial;
 			}
 			else
 			{
-				v = (v >> 1);
+				v = (UInt16)(v >> 1);
 			}
 		}
 		tab[i] = v;
@@ -27,21 +27,21 @@ extern "C" void CRC16R_InitTable(UInt16 *tab, UInt16 polynomial)
 	i = 256;
 	while (i-- > 0)
 	{
-		tab[256  + i] = (tab[0    + i] >> 8) ^ tab[tab[0    + i] & 0xff];
-		tab[512  + i] = (tab[256  + i] >> 8) ^ tab[tab[256  + i] & 0xff];
-		tab[768  + i] = (tab[512  + i] >> 8) ^ tab[tab[512  + i] & 0xff];
-		tab[1024 + i] = (tab[768  + i] >> 8) ^ tab[tab[768  + i] & 0xff];
-		tab[1280 + i] = (tab[1024 + i] >> 8) ^ tab[tab[1024 + i] & 0xff];
-		tab[1536 + i] = (tab[1280 + i] >> 8) ^ tab[tab[1280 + i] & 0xff];
-		tab[1792 + i] = (tab[1536 + i] >> 8) ^ tab[tab[1536 + i] & 0xff];
-		tab[2048 + i] = (tab[1792 + i] >> 8) ^ tab[tab[1792 + i] & 0xff];
-		tab[2304 + i] = (tab[2048 + i] >> 8) ^ tab[tab[2048 + i] & 0xff];
-		tab[2560 + i] = (tab[2304 + i] >> 8) ^ tab[tab[2304 + i] & 0xff];
-		tab[2816 + i] = (tab[2560 + i] >> 8) ^ tab[tab[2560 + i] & 0xff];
-		tab[3072 + i] = (tab[2816 + i] >> 8) ^ tab[tab[2816 + i] & 0xff];
-		tab[3328 + i] = (tab[3072 + i] >> 8) ^ tab[tab[3072 + i] & 0xff];
-		tab[3584 + i] = (tab[3328 + i] >> 8) ^ tab[tab[3328 + i] & 0xff];
-		tab[3840 + i] = (tab[3584 + i] >> 8) ^ tab[tab[3584 + i] & 0xff];
+		tab[256  + i] = (UInt16)(tab[0    + i] >> 8) ^ tab[tab[0    + i] & 0xff];
+		tab[512  + i] = (UInt16)(tab[256  + i] >> 8) ^ tab[tab[256  + i] & 0xff];
+		tab[768  + i] = (UInt16)(tab[512  + i] >> 8) ^ tab[tab[512  + i] & 0xff];
+		tab[1024 + i] = (UInt16)(tab[768  + i] >> 8) ^ tab[tab[768  + i] & 0xff];
+		tab[1280 + i] = (UInt16)(tab[1024 + i] >> 8) ^ tab[tab[1024 + i] & 0xff];
+		tab[1536 + i] = (UInt16)(tab[1280 + i] >> 8) ^ tab[tab[1280 + i] & 0xff];
+		tab[1792 + i] = (UInt16)(tab[1536 + i] >> 8) ^ tab[tab[1536 + i] & 0xff];
+		tab[2048 + i] = (UInt16)(tab[1792 + i] >> 8) ^ tab[tab[1792 + i] & 0xff];
+		tab[2304 + i] = (UInt16)(tab[2048 + i] >> 8) ^ tab[tab[2048 + i] & 0xff];
+		tab[2560 + i] = (UInt16)(tab[2304 + i] >> 8) ^ tab[tab[2304 + i] & 0xff];
+		tab[2816 + i] = (UInt16)(tab[2560 + i] >> 8) ^ tab[tab[2560 + i] & 0xff];
+		tab[3072 + i] = (UInt16)(tab[2816 + i] >> 8) ^ tab[tab[2816 + i] & 0xff];
+		tab[3328 + i] = (UInt16)(tab[3072 + i] >> 8) ^ tab[tab[3072 + i] & 0xff];
+		tab[3584 + i] = (UInt16)(tab[3328 + i] >> 8) ^ tab[tab[3328 + i] & 0xff];
+		tab[3840 + i] = (UInt16)(tab[3584 + i] >> 8) ^ tab[tab[3584 + i] & 0xff];
 	}
 }
 
@@ -73,12 +73,12 @@ extern "C" UInt16 CRC16R_Calc(const UInt8 *buff, UOSInt buffSize, UInt16 *tab, U
 	{
 		currVal ^= ReadUInt16(buff);
 		buff += 2;
-		currVal = tab[256 + (currVal & 0xff)] ^ tab[0 + (currVal >> 8)];
+		currVal = (UInt16)tab[256 + (currVal & 0xff)] ^ tab[0 + (currVal >> 8)];
 		buffSize -= 2;
 	}
 	while (buffSize--)
 	{
-		currVal = tab[(currVal & 0xff) ^ (*buff)] ^ (currVal >> 8);
+		currVal = tab[(currVal & 0xff) ^ (*buff)] ^ (UInt16)(currVal >> 8);
 		buff++;
 	}
